app: theme API, rename preload to appAPI and make settings page
* add a actual theme API
* rename preload script to appAPI
* make settings button open the settings page
* make settings page use the new UI style
